+++ com/authorize/http/response/modes/webmessage.js
@@ -2,5 +2,5 @@ exports = module.exports = function() {
   return require('oauth2orize-wmrm');
 };
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/authorization/http/ResponseMode';
+exports['@implements'] = 'module:oauth2orize.Responder';
 exports['@mode'] = 'web_message';