+++ com/authorize/http/response/types/code.js
@@ -8,7 +8,7 @@ exports = module.exports = function(acs, logger, C) {
       var modes = {};
       
       return new Promise(function(resolve, reject) {
-        var components = C.components('http://i.authnomicon.org/oauth2/authorization/http/ResponseMode')
+        var components = C.components('module:oauth2orize.Responder')
           , key;
       
         (function iter(i) {