+++ test/authorize/http/response/modes/webmessage.test.js
@@ -7,7 +7,7 @@ var factory = require('../../../../../com/authorize/http/response/modes/webmessa
 describe('authorize/http/response/modes/webmessage', function() {
   
   it('should be annotated', function() {
-    expect(factory['@implements']).to.equal('http://i.authnomicon.org/oauth2/authorization/http/ResponseMode');
+    expect(factory['@implements']).to.equal('module:oauth2orize.Responder');
     expect(factory['@mode']).to.equal('web_message');
     expect(factory['@singleton']).to.be.undefined;
   });